# apps/lms-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  service.CreateTestSeriesRequest:
    properties:
      description:
        minLength: 10
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      passing_percentage:
        maximum: 100
        minimum: 0
        type: number
      show_correct_answers:
        type: boolean
      shuffle_questions:
        type: boolean
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - description
    - duration_minutes
    - passing_percentage
    - title
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      role:
        enum:
        - student
        - faculty
        - admin
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
...
  service.TestSeriesListResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      is_published:
        type: boolean
      passing_percentage:
        type: number
      slug:
        type: string
      title:
        type: string
      total_questions:
        type: integer
    type: object
  service.TestSeriesResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      is_published:
        type: boolean
      passing_percentage:
        type: number
      show_correct_answers:
        type: boolean
      shuffle_questions:
        type: boolean
      slug:
        type: string
      title:
        type: string
      total_questions:
        type: integer
      updated_at:
        type: string
    type: object
...
  service.UpdateTestSeriesRequest:
    properties:
      description:
        minLength: 10
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      is_published:
        type: boolean
      passing_percentage:
        maximum: 100
        minimum: 0
        type: number
      show_correct_answers:
        type: boolean
      shuffle_questions:
        type: boolean
      title:
        maxLength: 255
        minLength: 3
        type: string
    type: object
  service.UpdateUserRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  service.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      last_name:
        type: string
      phone:
        type: string
      profile_picture_url:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Learning Management System Backend API with Course, Test, and Student
    Management
  license:
    name: MIT
  title: LMS Backend API
  version: "1.0"
paths:
...
  /test-series:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of test series with optional filters
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of tests per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Filter by status (draft, published)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.TestSeriesListResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List all tests
      tags:
      - Test Series
    post:
      consumes:
      - application/json
      description: Create a new test/quiz (faculty/admin only)
      parameters:
      - description: Test series creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateTestSeriesRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TestSeriesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a new test series
      tags:
      - Test Series
  /test-series/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing test series (admin only)
      parameters:
      - description: Test Series UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a test series
      tags:
      - Test Series
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific test series
      parameters:
      - description: Test Series UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TestSeriesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get test series details
      tags:
      - Test Series
    patch:
      consumes:
      - application/json
      description: Update an existing test series (faculty/admin only)
      parameters:
      - description: Test Series UUID
        in: path
        name: id
        required: true
        type: string
      - description: Test series update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateTestSeriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TestSeriesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update test series details
      tags:
      - Test Series
...
  /test-series/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publish a draft test series (faculty/admin only)
      parameters:
      - description: Test Series UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TestSeriesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Publish a test series
      tags:
      - Test Series
...
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of all users (admin only)
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of users per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Filter by role (student, faculty, admin)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account (admin only)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific user
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get user details
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user details (user can update own profile, admin can update
        any profile)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: User update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update user information
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Add a JWT token to the header with ** Bearer &lt;JWT&gt; ** token
      to authorize"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"